Fixed some semaphore problems but apparently not all
