+++ src/DevonianModel.ts
@@ -1,0 +1,6 @@
+
+import { IdentifierMap } from './IdentifierMap.js';
+
+export type DevonianModel = {
+  foreignIds: IdentifierMap;
+};
