+++ src/IdentifierMap.ts
@@ -1,0 +1,3 @@
+export type IdentifierMap = {
+  [platform: string]: string | number;
+};
